Remove ClientSecretCredential, EnvironmentCredential
Also moves SpecificAzureCredential to a sample. Fixes #1812.
