+++ sdk/identity/azure_identity/src/credentials/mod.rs
@@ -14,12 +14,9 @@ mod azure_cli_credentials;
 mod cache;
 #[cfg(feature = "client_certificate")]
 mod client_certificate_credentials;
-mod client_secret_credentials;
 mod default_credentials;
-mod environment_credentials;
 mod imds_managed_identity_credentials;
 mod options;
-mod specific_azure_credential;
 mod virtual_machine_managed_identity_credential;
 mod workload_identity_credentials;
 
@@ -28,11 +25,9 @@ pub use app_service_managed_identity_credential::*;
 pub use azure_cli_credentials::*;
 #[cfg(feature = "client_certificate")]
 pub use client_certificate_credentials::*;
-pub use client_secret_credentials::*;
 pub use default_credentials::*;
-pub use environment_credentials::*;
+pub use imds_managed_identity_credentials::ImdsId;
 pub(crate) use imds_managed_identity_credentials::*;
 pub use options::*;
-pub use specific_azure_credential::*;
 pub use virtual_machine_managed_identity_credential::*;
 pub use workload_identity_credentials::*;
